(PC-33723)[API] chore: move AdageFrontRoles to core/educational/models.py
