+++ api/src/pcapi/routes/adage_iframe/serialization/adage_authentication.py
@@ -1,4 +1,3 @@
-import enum
 import logging
 
 from pydantic.v1 import Field
@@ -6,6 +5,7 @@ from pydantic.v1 import ValidationError
 
 from pcapi.core.educational import models as educational_models
 from pcapi.core.educational.exceptions import MissingRequiredRedactorInformation
+from pcapi.core.educational.models import AdageFrontRoles
 from pcapi.core.educational.schemas import RedactorInformation
 from pcapi.routes.serialization import BaseModel
 from pcapi.serialization.utils import to_camel
@@ -14,11 +14,6 @@ from .redactor import RedactorPreferences
 
 
 logger = logging.getLogger(__name__)
-
-
-class AdageFrontRoles(enum.Enum):
-    REDACTOR = "redactor"
-    READONLY = "readonly"
 
 
 class AuthenticatedInformation(BaseModel):
