+++ api/tests/routes/adage_iframe/offers_test.py
@@ -5,9 +5,9 @@ import pytest
 
 from pcapi.core.educational import factories as educational_factories
 from pcapi.core.educational import models as educational_models
+from pcapi.core.educational.models import AdageFrontRoles
 import pcapi.core.educational.testing as adage_api_testing
 import pcapi.core.educational.utils as educational_utils
-from pcapi.routes.adage_iframe.serialization.adage_authentication import AdageFrontRoles
 
 
 pytestmark = pytest.mark.usefixtures("db_session")
