+++ api/tests/routes/adage_iframe/post_logs_test.py
@@ -4,8 +4,8 @@ from flask import url_for
 import pytest
 
 from pcapi.core.educational import utils
+from pcapi.core.educational.models import AdageFrontRoles
 from pcapi.core.educational.utils import get_hashed_user_id
-from pcapi.routes.adage_iframe.serialization.adage_authentication import AdageFrontRoles
 from pcapi.routes.adage_iframe.serialization.logs import PaginationType
